Separate disassembler
Separate disassembler code into its own file. The code is long enough to deserve
it.
