+++ disassemble.h
@@ -1,0 +1,1 @@
+void disassemble(unsigned op);
